+++ indic_transliteration/sanscript/__init__.py
@@ -58,8 +58,6 @@ from __future__ import unicode_literals
 # ---------------
 #: Internal name of Bengali. Bengali ``ba`` and ``va`` are both rendered
 #: as `ব`.
-import sys
-
 from indic_transliteration.sanscript.schemes import Scheme
 from indic_transliteration.sanscript.schemes import roman
 from indic_transliteration.sanscript.schemes.brahmic import northern, southern, eastern
@@ -88,6 +86,13 @@ KOLKATA = roman.KOLKATA
 SLP1 = roman.SLP1
 VELTHUIS = roman.VELTHUIS
 WX = roman.WX
+
+## NOTE: See the Scheme constructor documentation for a few general notes while defining schemes.
+SCHEMES = {}
+SCHEMES.update(roman.SCHEMES)
+SCHEMES.update(northern.SCHEMES)
+SCHEMES.update(southern.SCHEMES)
+SCHEMES.update(eastern.SCHEMES)
 
 
 class SchemeMap(object):
@@ -183,188 +188,6 @@ class SchemeMap(object):
                            "virama":  self.virama,
                            "consonants": self.consonants})
 
-def _roman(data, scheme_map, **kw):
-  """Transliterate `data` with the given `scheme_map`. This function is used
-  when the source scheme is a Roman scheme.
-
-  :param data: the data to transliterate
-  :param scheme_map: a dict that maps between characters in the old scheme
-                     and characters in the new scheme
-  """
-  vowels = scheme_map.vowels
-  marks = scheme_map.marks
-  virama = scheme_map.virama
-  consonants = scheme_map.consonants
-  non_marks_viraama = scheme_map.non_marks_viraama
-  max_key_length_from_scheme = scheme_map.max_key_length_from_scheme
-  to_roman = scheme_map.to_scheme.is_roman
-
-  togglers = kw.pop('togglers', set())
-  suspend_on = kw.pop('suspend_on', set())
-  suspend_off = kw.pop('suspend_off', set())
-  if kw:
-    raise TypeError('Unexpected keyword argument %s' % list(kw.keys())[0])
-
-  buf = []
-  i = 0
-  had_consonant = found = False
-  len_data = len(data)
-  append = buf.append
-
-  # If true, don't transliterate. The toggle token is discarded.
-  toggled = False
-  # If true, don't transliterate. The suspend token is retained.
-  # `suspended` overrides `toggled`.
-  suspended = False
-
-  while i <= len_data:
-    # The longest token in the source scheme has length `max_key_length_from_scheme`. Iterate
-    # over `data` while taking `max_key_length_from_scheme` characters at a time. If we don`t
-    # find the character group in our scheme map, lop off a character and
-    # try again.
-    #
-    # If we've finished reading through `data`, then `token` will be empty
-    # and the loop below will be skipped.
-    token = data[i:i + max_key_length_from_scheme]
-
-    while token:
-      if token in togglers:
-        toggled = not toggled
-        i += 2  # skip over the token
-        found = True  # force the token to fill up again
-        break
-
-      if token in suspend_on:
-        suspended = True
-      elif token in suspend_off:
-        suspended = False
-
-      if toggled or suspended:
-        token = token[:-1]
-        continue
-
-      # Catch the pattern CV, where C is a consonant and V is a vowel.
-      # V should be rendered as a vowel mark, a.k.a. a "dependent"
-      # vowel. But due to the nature of Brahmic scripts, 'a' is implicit
-      # and has no vowel mark. If we see 'a', add nothing.
-      if had_consonant and token in vowels:
-        mark = marks.get(token, '')
-        if mark:
-          append(mark)
-        elif to_roman:
-          append(vowels[token])
-        found = True
-
-      # Catch any non_marks_viraama character, including consonants, punctuation,
-      # and regular vowels. Due to the implicit 'a', we must explicitly
-      # end any lingering consonants before we can handle the current
-      # token.
-      elif token in non_marks_viraama:
-        if had_consonant:
-          append(virama[''])
-        append(non_marks_viraama[token])
-        found = True
-
-      if found:
-        had_consonant = token in consonants
-        i += len(token)
-        break
-      else:
-        token = token[:-1]
-
-    # We've exhausted the token; this must be some other character. Due to
-    # the implicit 'a', we must explicitly end any lingering consonants
-    # before we can handle the current token.
-    if not found:
-      if had_consonant:
-        append(virama[''])
-      if i < len_data:
-        append(data[i])
-        had_consonant = False
-      i += 1
-
-    found = False
-
-  return ''.join(buf)
-
-
-def _brahmic(data, scheme_map, **kw):
-  """Transliterate `data` with the given `scheme_map`. This function is used
-  when the source scheme is a Brahmic scheme.
-
-  :param data: the data to transliterate
-  :param scheme_map: a dict that maps between characters in the old scheme
-                     and characters in the new scheme
-  """
-  marks = scheme_map.marks
-  virama = scheme_map.virama
-  consonants = scheme_map.consonants
-  non_marks_viraama = scheme_map.non_marks_viraama
-  to_roman = scheme_map.to_scheme.is_roman
-  max_key_length_from_scheme = scheme_map.max_key_length_from_scheme
-
-  buf = []
-  i = 0
-  to_roman_had_consonant = found = False
-  append = buf.append
-  # logging.debug(pprint.pformat(scheme_map.consonants))
-
-  # We dont just translate each brAhmic character one after another in order to prefer concise transliterations when possible - for example ज्ञ -> jn in optitrans rather than j~n.
-  while i <= len(data):
-    # The longest token in the source scheme has length `max_key_length_from_scheme`. Iterate
-    # over `data` while taking `max_key_length_from_scheme` characters at a time. If we don`t
-    # find the character group in our scheme map, lop off a character and
-    # try again.
-    #
-    # If we've finished reading through `data`, then `token` will be empty
-    # and the loop below will be skipped.
-    token = data[i:i + max_key_length_from_scheme]
-
-    while token:
-      if len(token) == 1:
-        if token in marks:
-          append(marks[token])
-          found = True
-        elif token in virama:
-          append(virama[token])
-          found = True
-        else:
-          if to_roman_had_consonant:
-            append('a')
-          append(non_marks_viraama.get(token, token))
-          found = True
-      else:
-        if token in non_marks_viraama:
-          if to_roman_had_consonant:
-            append('a')
-          append(non_marks_viraama.get(token))
-          found = True
-
-      if found:
-        to_roman_had_consonant = to_roman and token in consonants
-        i += len(token)
-        break        
-      else:
-        token = token[:-1]
-
-    # Continuing the outer while loop.
-    # We've exhausted the token; this must be some other character. Due to
-    # the implicit 'a', we must explicitly end any lingering consonants
-    # before we can handle the current token.
-    if not found:
-      if to_roman_had_consonant:
-        append(next(iter(virama.values())))
-      if i < len(data):
-        append(data[i])
-        to_roman_had_consonant = False
-      i += 1
-
-    found = False
-
-  if to_roman_had_consonant:
-    append('a')
-  return ''.join(buf)
-
 
 @lru_cache(maxsize=8)
 def _get_scheme_map(input_encoding, output_encoding):
@@ -403,17 +226,11 @@ def transliterate(data, _from=None, _to=None, scheme_map=None, **kw):
   }
   options.update(kw)
 
+  from indic_transliteration.sanscript.brahmic_mapper import _brahmic
+  from indic_transliteration.sanscript.roman_mapper import _roman
   func = _roman if scheme_map.from_scheme.is_roman else _brahmic
   return func(data, scheme_map, **options)
 
 
 def get_standard_form(data, scheme_name):
   return transliterate(data=transliterate(data=data, _from=scheme_name, _to=DEVANAGARI), _from=DEVANAGARI, _to=scheme_name)
-
-
-## NOTE: See the Scheme constructor documentation for a few general notes while defining schemes.
-SCHEMES = {}
-SCHEMES.update(roman.SCHEMES)
-SCHEMES.update(northern.SCHEMES)
-SCHEMES.update(southern.SCHEMES)
-SCHEMES.update(eastern.SCHEMES)
