+++ setup.py
@@ -28,7 +28,7 @@ setup(
   # Versions should comply with PEP440.  For a discussion on single-sourcing
   # the version across setup.py and the project code, see
   # https://packaging.python.org/en/latest/single_source_version.html
-  version='1.8.6',
+  version='1.8.7',
 
   description='Transliteration tools to convert text in one indic script encoding to another',
   long_description=long_description,
