+++ tests/data/transliterationTests.json
@@ -91,8 +91,8 @@
   "to_devanaagarii": [
     {
       "description": "Check aDDak and tippee",
-      "dev": "अब्भर अक्क अभंग भुल्ला",
-      "gurmukhi": "ਅੱਭਰ ਅਕੱ ਅਭੰਗ ਭੁਲਾੱ"
+      "dev": "अब्भर अक्क मूंडा भुल्ला",
+      "gurmukhi": "ਅੱਭਰ ਅੱਕ ਮੂੰਡਾ ਭੁੱਲਾ"
     },
     {
       "description": "Tests Capitalized IAST",
